fix: upsert summary race condition
Doing a first query to calculate which records will be inserted/updated
and then upserting them could lead to a race condition where multiple
concurrent executions would assume there are about to insert the same
record but only the first one to execute the second query is actually
doing the insertion. The upsert summary 'added' value would then be wrong,
effectively counting the same records multiple times. This is then
affecting the records_count table with a saved count higher than the
actual count
It is unlikely to happen unless `batchSave` is executed concurrently
which we are not recommending but in theory it does happen