fix: upsert summary race condition (#2952)
Doing a first query to calculate which records will be inserted/updated
and then upserting them could lead to a race condition when multiple
concurrent executions would assume there are about to insert the same
record but only the first one to execute the second query is actually
doing the insertion. The upsert summary 'added' value would then be
wrong, effectively counting the same records multiple times. This is
then affecting the records_count table with a saved count higher than
the actual count
It is unlikely to happen unless `batchSave` is executed concurrently
which we are not recommending but in theory it does happen since there
is about a dozen rows in records_count table that are slightly higher
than the real counts.

The fix consist in adding a advisory lock to prevent concurrent upsert
by the same sync job and a single CTE query comparing before/after to
determine which records are inserted/updated/deleted/undeleted

## Issue ticket number and link
Required to fix discrepancies between real records count and the
records_count table

https://linear.app/nango/issue/NAN-1929/[trial]-show-object-count-in-sync-tab

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [x] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: