feat: rename post-connection-scripts table to on-event-scripts
Nango currently allows users to define a custom script that would be
executed when a connection is created.
We want to extend this feature to also make it possible to execute
a custom script when a connection is deleted.
In order to do so we are going to implement a generic solution where a
custom script can be executed based on a Nango event like connection
creation or deletion
This first PR is not introducing any new behaviour but rename the
post-connection-scripts into on-event-scripts and adding a event column
where the type of event associated with the script is being stored.
The PR is also renaming things related to the deployment, making sure
the old `postConnectionScriptsByProvider` param sent by old CLI is still
supported for backward compatibility

In next PRs I will:
- rename all things related to postConnnection in the orchestrator/scheduler
- add support for ON_CONNECTION_DELETED event
- refactor CLI to allow definition of both events/scripts