chore(orchestrator): rename PostConnectionTask to OnEventTask
No behavior is changing as part of this PR. It is just renaming the
PostConnectionTask into OnEventTask.
We could find an even more generic name because from the point of view
of the orchestrator it is just a task with a script file attached to it
but I couldn't find a better name. I am open to suggestion though.
There is still a post connection concept now. It is the provider-specific
piece of logic that is define in providers.yaml (ex: https://github.com/NangoHQ/nango/blob/master/packages/shared/providers.yaml#L1080)
I don't like the fact it is called post_connection_script since it is
not a script but some prebuilt logic that is executed server side.
Anyway renaming it is out of scope for this PR.