chore(orchestrator): rename PostConnectionTask to OnEventTask (#2997)
## Describe your changes

No behavior is changing as part of this PR. It is just renaming the
PostConnectionTask into OnEventTask and tighten some types.
We could find an even more generic name because, from the point of view
of the orchestrator, it is just a task with a script file attached to it
but I couldn't find a better name. I am open to suggestion though.

The concept of post connection still exists in the codebase. It is the
provider-specific piece of logic that is defined in providers.yaml (ex:
https://github.com/NangoHQ/nango/blob/master/packages/shared/providers.yaml#L1080)
I don't like the fact it is called post_connection_script since it is
not a script but some prebuilt logic that is executed server side.
Anyway renaming it is out of the scope of this PR but at least now we
don't have 2 different things with the same name anymore.

Next PR will refactor the CLI to enable the definition of the events and
their associated scripts

## Issue ticket number and link

https://linear.app/nango/issue/NAN-1962/connect-lifecycle-scripts#comment-7a350835

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: