fix(runner): do not exit the runner process when a sync is cancelled
I think that at some point we were forking/spawning a new subprocess for
each script execution. When aborting we will then exit the process.
However we don't do that anymore and scripts simply run in a node vm in
the same runner process.
This commit modifies the sync sdk so an error is thrown instead of
exiting. Jobs then discard this specific error.