fix(runner): do not exit the runner process when a sync is cancelled (#3008)
## Describe your changes

I think that at some point we were forking/spawning a new subprocess for
each script execution. When aborting we will then exit the process.
However we don't do that anymore and scripts simply run in a node vm in
the same runner process.
This commit modifies the sync sdk so an error is thrown instead of
exiting. Jobs then discard this specific error.

## Issue ticket number and link

https://linear.app/nango/issue/NAN-2195/fix-aborting-a-script-kills-the-runner-process

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: