fix(webapp): Make sure integrations list is in sync with connections (#3006)
## Describe your changes

Previously we weren't revalidating data relating to integrations in SWR
when connections were added and deleted, which made the integrations
list feel stale. This fixes that.

Additionally it adds a 15 second refresh on the integrations data in SWR
so that connections that are created outside of the dashboard will still
appear in time and give the customers a better feeling that things are
happening.

In the future it would probably be ideal to trigger this with a
websocket instead of an interval, but this works for now!

## Issue ticket number and link


https://linear.app/nango/issue/NAN-1310/caching-is-too-aggressive-on-the-app

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: