fix(webapp): Clear any cache from swr/infinite (#3005)
## Describe your changes

We were seeing the connections tab still populated with stale data after
logging out and logging back in. The culprit was that swr/infinite does
not clear its state through the same mechanism as swr itself. This adds
code to work around that, and a brute force page reload to clear all
state in memory.

## Issue ticket number and link


https://linear.app/nango/issue/NAN-1310/caching-is-too-aggressive-on-the-app

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: