feat: add support for on-events syntax in nango.yaml
We want to be able to support custom script for more than just
post-connection event. This commit is adding support for the following
syntax in nango.yaml:
```
my-integration:
    on-events:
        post-connection-creation:
            - script1
            - script2
        pre-connection-deletion:
            - script1
    syncs:
        ...
```
The change is backward compatible and still support the old
`post-connection-scripts` even though I don't think it is currently
being used.
I will add in next PR the logic to trigger the script when connection is
deleted