chore: update pull_request_template
My take: having a checklist of things that rarely applies is pretty
useless. I (and I am sure not the only one) just ignore the
observability and analytics points. I propose this change to focus on
the things that matter: description of the change, testing and ticket
All of them almost always applies.

Also changing to comments so the template instructions don't show once
the PR is submitted