chore: update pull_request_template (#3033)
My take: having a checklist of things that rarely apply is pretty
useless. I (and I am sure not the only one) just always ignore the
observability and analytics points. I propose this change to focus on
the things that matter: description of the changes, testing and ticket.
All of them almost always applies.

Also changing to comments so the template instructions don't show once
the PR is submitted

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: not applicable 🤔 
- [ ] I added observability, otherwise the reason is: not applicable 🙄 
- [ ] I added analytics, otherwise the reason is: not applicable 😝