fix(webapp, server): fix custom auth issues (#3042)
I bumped into a few spots where custom auth code wasn't making sense to
me, and after further digging think this is the solution. 🫨

Here we were reusing the oauth client id on two fields:


https://github.com/NangoHQ/nango/blob/173d6be59837a53c4c3991f72a4f7fa2c44a3cfe/packages/webapp/src/pages/Integrations/providerConfigKey/Settings/components/Custom.tsx#L24

And here we were setting client id from two fields:


https://github.com/NangoHQ/nango/blob/716cd13b824b5160add1cacee2438717bf36c855/packages/server/lib/controllers/v1/integrations/providerConfigKey/patchIntegration.ts#L109-L112

Ultimately I realized that we're using app_id from the custom field in
most of the logic for that provider:


https://github.com/NangoHQ/nango/blob/204f94bb3d6ef927540bf985e124b7989587d08e/packages/server/lib/controllers/config.controller.ts#L224-L229


<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

## How I tested it:

This was convoluted!

- Set up a new github app (go to settings, dev settings, create app)
- Be sure to check `Request user authorization (OAuth) during
installation` as part of the setup
- Set up a new integration for `github-app-oauth`
- Fill in all the settings
- Click add connection (headless) and create the connection
- With this integration, if you somehow already have it installed you'll
end up on the installation page and it's not apparent in the small
connect window.