fix(runner): Remove setInterval from providers reload (#3057)
<!-- Describe the problem and your solution --> 
`setInterval` can cause problems because it doesn't ensure that multiple
callbacks don't run at the same time. This removes `setInterval` from
provider reloading.

Note: I'm not sure if it's worth it in this case to deploy runners or
just wait until we deploy them next time.

<!-- Issue ticket number and link (if applicable) -->

See https://linear.app/nango/issue/NAN-2245/fix-do-not-use-setinterval



<!-- Testing instructions (skip if just adding/editing providers) -->


## How I tested it:

- Make sure a runner has started
- Edit the providers.yaml
- Wait for the reloadInterval (60s by default) to ensure you get a
message that it was updated