feat(server): Add missing fields column for integrations to db (#3045)
<!-- Describe the problem and your solution --> 
We need a stored set of issues on integrations so that we can show those
issues in the UI. This will add the new column and ensure that it's
handled appropriately in the app.

- [x] Add new column
- [x] Set up validation function for integrations
- [x] Update creating integrations to fill in fields
- [x] Update updating integrations to fill in fields

A future PR will migrate existing providers to note their missing fields
correctly since, and then we'll be able to utilize those in the
dashboard.

<!-- Issue ticket number and link (if applicable) -->


https://linear.app/nango/issue/NAN-2168/surface-integrationsconnections-errors-in-nango-ui

<!-- Testing instructions (skip if just adding/editing providers) -->
## How I Tested It
- Run the tests
- Click through adding a new oauth based integration
- Note that missing_fields is set for it in the DB
- Update that integration
- Note that missing_fields has been updated accordingly