fix(authorization): authorization url encoding (#3062)
### Describe the problem and your solution
Turns out some of the providers don't like it when we encode some of the
url parts, i.e `scopes`, `base_url` when we are interpolating with
`connectionConfigs`. To help with this, the PR introduces a way we can
handle this within the various parts. Currently we can skip
authorization_url encoding within the `scopes`, or when interpolating
the `base_url` with `connectionConfig`.
<!-- Issue ticket number and link (if applicable) -->

### Testing instructions
When creating a new connection for `Pennylane`, the scopes aren't
encoded `customer_invoices+supplier_invoices+ledger+accounting`

---------

Co-authored-by: Hassan Wari <hassanwari@Hassans-MacBook-Pro.local>