feat(server): Backfill missing_fields on _nango_config (#3050)
<!-- Describe the problem and your solution --> 
Fills out `missing_fields` as best as we can based on `_nango_config`.
Unfortunately the `CUSTOM` `auth_method` won't be able to be backfilled
performantly, but all others were possible.

For all of these, the prior PR will cause them to show errors in the
future when created or edited, so I think we're safe here without having
to decrypt all of the `CUSTOM` auth based ones.

<!-- Issue ticket number and link (if applicable) -->


https://linear.app/nango/issue/NAN-2168/surface-integrationsconnections-errors-in-nango-ui

<!-- Testing instructions (skip if just adding/editing providers) -->

## How I tested

- Run the migration against dev db with oauth and app auth_method
integrations that don't have their settings filled and don't have errors
set yet. Feel free to just clear `missing_fields` in the DB as an easy
way to handle that.
- Confirm in the database that the fields were correctly filled after
the migration runs.