feat(api): expose credentials in GET /integrations/:uniqueKey (#3074)
## Changes

- `GET /integrations/:uniqueKey` now expose credentials (optional)
Some customers need the credentials for internal operations or oauth
revoke. This expose optionally the credentials with some added security.

- Try to fix IntegrationConfig type by creating a smaller scope for
integration create
This type is annoying, it's used a bit everywhere and everything is
marked as undefined except it's null not undefined. Anyway, first steps
since I needed better types.