ci: split CLI/Node tests  (#3096)
## Changes

Fixes
https://linear.app/nango/issue/NAN-2181/ci-split-all-cli-node-client-out-of-main-tests-suite

- Split CLI and node client tests and execute in a dedicated matrix 
They run 6 tests (Node 18, 20, 22 x windows or ubuntu). It's a lot but
seems relatively fast. Alternatively we could make windows matrix only
execute on master.
We could further optimize this and run them only types or node or cli
has changed but it's a pain to create this kind of logic.

- Integration and regular tests only run on Ubuntu + .nvmrc
This doesn't save much time since the matrix was running in parallel but
reduce potential for flaky tests