feat: fleet into jobs/server (#3129)
I apologize in advance for the big PR.
Fleet is now integrated in jobs and server. It is being a feature flag
because it only support local runners for now. (The migration creating
the tables would run in staging/prod though)

This PR:
- adds runner routing logic in jobs (behind flag)
- implement local runner logic (pretty much copied from previous logic)
- adds register and idle mechanism/endpoints between jobs and runners
- fixes bugs with the fleet supervisor 

Note: I want fleet to manage local runner as well so it is the same code
(almost) running locally and in prod. In order to do so every time jobs
starts locally a new fake deployment is added to the db so fleet
automatically rolls out a new runner and manage its lifecycle. Other
than that there are tiny adjustments of the logic that only exist for
local execution (ex: shorter timeouts, finishing timeout = idle)

What's not in this PR but will come next:
- /deploy endpoint 
- render node provider
- locking mechanism to ensure only one supervisor is running
- CI script to call /deploy endpoint
- runner config override mechanism for customers that must get more
runner resources

# How to test
1. enable feature flag locally
 a. `redis-cli`
 b. `set flag:fleet:global true`
2. start nango locally