feat(fleet): add node config overrides
Some customers are assigned a bigger runner instance.
This commit adds a node_config_overrides table that allows the fleet
supervisor to override the default node config if an entry exists in
this table for a given routingId
For now there is no tooling to add/update/remove entries in this table.
At first we'll update it by hand.