fix(runner): set render PORT (#3165)
https://render.com/docs/web-services#port-binding
Render default port for new service is 10000 and at startup they detect
any app running on another port and switch the primary port to it.
However the switch triggers a restart to reconfigure the network which
is problematic for us since the runner has already registered as RUNNING
and is being potentially sent task
This commit set the PORT explicitely in order to avoid the automatic
detection and restart

I have tested in staging. No awkward restart when setting the PORT
explicitely