feat(server): Add instance identifier when logging in cloud (#3166)
<!-- Describe the problem and your solution --> 
When reviewing logs in Render right now it's difficult to know what
instance of server logged what. This adds a random identifier at startup
that will help us filter logs to specific instances.

I thought about using hostnames but the hostnames on render are super
long and not that helpful, and I'm not 100% sure they're guaranteed to
be unique on their infra.

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->