feat(integrations): add support for lastpass (#3171)
## Describe the problem and your solution
Add support for LastPass
<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

There is no need to have the `proxy.verification` method since the
provider returns success 200 even when the verification is wrong.

Nango will supply header `Authorizaton: Basic
(username:password).toString('base64')` since it is of `BASIC` auth_mode
but seems the provider ingores this so we don't need to worry about it.

Co-authored-by: Hassan Wari <hassanwari@Hassans-MacBook-Pro.local>