fix: move triggering actions out of transations
When removing failing connection in SlackService, we used to trigger
actions in the middle of a db transaction that is locking rows.
We also do all sort of async operations like logging.
If the actions for instance are experiencing any kind of issue, the transaction is kept
open and waiting until the actions times out or the db cancel the
transaction.