chore: remove sentry (#3259)
## Changes

Fixes https://linear.app/nango/issue/NAN-2111/remove-sentry

- Remove sentry from backend
While it's good to have that, the current implementation was not useful.
We might want to reintroduce it later more intentionally.
- Frontend use /env.js for Sentry (and PostHog) so there is no need to
compile a different image per environment
This docker image was not used yet so it's mostly reducing noise and CI
time