chore: Fix up handler type (#3282)
<!-- Describe the problem and your solution --> 
I kept noticing spots where I was getting eslint errors because of our
handlers being async but the handler type expected `void` and not
`Promise<void>`. This updates the type for handlers and cleans up the
resulting type errors.

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->
If you want to see this in action, before switching to this branch check
out line 125 in `packages/jobs/lib/routes/tasks/putTask.ts`. It has an
eslint warning because of a misused promise. After this update that
eslint warning is gone.