fix(docker): stop building staging, prod, enterprise in favor of unified (#3281)
## Changes

- Stop building staging, prod, enterprise in favor of unified 
This means the unified `dockerfile` is built only once and not twice,
and I have removed enterprise staging and prod for server (aka we only
build one server).
After this is merged, the images won't be tagged by their env, it will
also save time on the CI and we might be able to deploy prod directly
from PR (handy for hot fixes but I need to test it first).

We still need to build `server-hosted` separately until we officially
support this use case. This already works but needs official
communication and deprecation period.