fix: fix log copy when setting frequency on script deployment
Log is currently stating `Sync frequency updated to X` on script deployment.
Since the frequency update doesn't necessary mean the value has changed
this commit is changing the log copy to say `Sync frequency is X` to be
less confusing.
We agreed with Khaliq about this simple change instead of changing the
internal API to differentiate between a frequency update resulting in a new
value instead and a noop one when value is the same