fix(http): use keepAlive and agent (almost) everywhere (#3318)
## Changes 

Contributes to
https://linear.app/nango/issue/SAM-6/backend-have-a-unified-fetcher-with-http-agent

- Use keepAlive and HTTP agent  (almost) everywhere
I started doing that months ago but never got to enable keepAlive (which
is not related to header but a way to keep a TCP connection open), since
we are doing mostly the same calls over and over again it should help
increase throughput and reduce network errors.

NB: some part of the code is using `fetch` which is not compatible with
this "old" way of doing things. Undici is exposing a new kind of Agent
that is specific.

NB2: For the ref, that's what I did previously at Algolia with very good
success, not sure it will have that much impact here but worth the try
https://www.algolia.com/blog/engineering/30-days-to-improve-our-crawler-performance-by-50-percent#:~:text=After%20all%20those%20fixes%20were%20deployed