fix(runner): tweak runner idling retry logic (#3324)
Not really a bug but the runner currently keeps notifying `jobs` that it
is idle every 10s.
However 10s might not be enough for jobs/fleet to terminate the runner
(if we are in the middle of a deploy for example, fleet might need to
start/terminate dozen of services and it can take quite a long time).
Following requests to /idle then fail with an error since the runner is
already considered idled. This commit increase the timeout when /idle
was called successfully to give enough time to jobs/fleet to terminate
the runner and avoid logging an error.
In theory we could also simply stop the loop once /idle has been
successful