feat(records): add upserting/updating records merging strategy
Currently, all entries saved/updated by a script override any
existing record.
We want to allow scripts to be able to define a merging strategy where
records that have been modified during the execution of the script
should not be overwritten.
To allow such behavior, this commit adds a parameter to the
Records.upsert and Records.update function to pass a cursor. Any entry
that would overwrite an existing records modified after this cursor will
be ignored instead of overwritting the records.