feat(records): add upserting/updating records merging strategy (#3379)
Currently, all entries saved/updated by a script override existing
records.
We want to allow scripts to be able to define a merging strategy where
records that have been modified during the execution of the script would
not be overwritten.
To allow such behavior, this commit adds a parameter to the
`Records.upsert` and `Records.update` function to pass a cursor. Any
entry that would overwrite an existing record modified after this cursor
will be ignored instead of overwriting the record.

# How to test
It is not wired yet so we'll have to rely on the tests