feat: add support for merging strategy into the runner sdk (#3420)
- adding `setMergingStrategy` runnder-sdk function
- if strategy is `ignore-if-modified-since` the function calls persist
`/cursor` and store it in a private variable. Its value is then
forwarded when calling `batchSave/Update/Delete` which are now sending
the next cursor. Again the updated cursor is forwarded to batchSave...
- One limitation of this approach is that it makes concurrent batchSave
undefined behavior. We already recommend not to do that anyway because
it doesn't play well with rate-limiting
- refactor persist API related code in runner-sdk into its own file
- send merging strategy to persist and handle response with next merging
strategy with new cursor
- fix a couple of bugs from previous PR
- modify CLI compiler to disallow setting merging strategy multiple
times or setting it after bachSave