chore(fleet): simplify node config overrides
NodeConfigOverride values are now either null (ie: don't override) or a given value. It simplifies adding new override since we don't have to worry about what are the default values, just enter null for the attributes that don't need to be overriden.

I still need to remove the commit_id column in deployment which won't be
used anymore once this commit is deployed. Will do it in next PR.