fix(cron): automatic refresh token was not passing decrypted connection (#3472)
## Changes 

Noticed something weird while improving the type for Connection in #3470

- Automatic refresh token was not passing decrypted connection


When doing this refactor https://github.com/NangoHQ/nango/pull/2870 I
split the refresh and the getConnection because it was inefficient. I
forgot to decrypt the connection, and due to bad typing, there was no
difference between Connection and (decrypted)Connection. The really
permissive code completely allowed this to pass, and still mark
connection has refreshed while not being refreshed.

NB: this was not impacting the regular refresh on fetch