fix(types): Connection -> DBConnection (#3470)
## Changes

> [!NOTE]
> Lot of files have been modified but it's mostly due to setting `id` as
not undefined
> 98% of this PR is type change only which can be scary but not as
serious as it seems

- **Move connection types** to `types` package
- **Replaced** `Connection` -> `DBConnection`
Fixed all undefined with nullable or not undefined
- **Added** `DBConnectionDecrypted` to differentiate between raw
connection and decrypted connection
This difference used to be marked by Connection vs StoredConnection
which was not very clear, not used correctly everywhere and was not
incompatible so it lead to bad argument being passed

- **Replaced** NangoConnection by ConnectionJobs and ConnectionInternal
`ConnectionJobs` is used by Jobs (yes) and `ConnectionInternal` is some
kind of weird shape that is used somewhere. Ideally this last type
should disappear

## 🤓 How to review?

You will see most changes are super trivial but if you want to start
with the right files:
-
[types/connection/db.ts](https://github.com/NangoHQ/nango/pull/3470/files#diff-a819f17557dd02b45e15723a24aba70c7847c50d6c4fedb928de1e9b8c0ae04f)
-
[connection.service.ts](https://github.com/NangoHQ/nango/pull/3470/files#diff-00f759b1e40c244e1f3258270374f4fb3e6ddf4e13ced14e12794ff30131efdd)
- then everything else


## 🧪 Tests

While it's only types change, it's also non-trivial. I would be very
glad to have more local tests
- Launch the UI and tests all pages
- Trigger some syncs and actions
- Use the local CLI to dryrun and deploy